requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.27.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
djangorestframework-simplejwt==5.4.0
sqlparse==0.5.3
typing_extensions==4.15.0

# OCR dependencies
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0

# HTTP client for KRA API
requests==2.31.0
